requirements.txt
Django==1.6.1
Pillow==2.2.2
distribute==0.6.24
dj-database-url==0.2.2
djangorestframework==2.3.10
factory-boy==2.3.0
simplejson==3.3.1
